Fix type narrowing warnings for MSVC

+++ gtn/graph.h
@@ -105,7 +105,7 @@ class Graph {
    * @param label The arc label.
    * @return The id of the added arc.
    */
-  int addArc(int srcNode, int dstNode, int label);
+  size_t addArc(size_t srcNode, size_t dstNode, int label);
 
   /**
    * Add a arc between two nodes.
@@ -116,9 +116,9 @@ class Graph {
    * @param weight The arc weight.
    * @return The id of the added arc.
    */
-  int addArc(
-      int srcNode,
-      int dstNode,
+  size_t addArc(
+      size_t srcNode,
+      size_t dstNode,
       int ilabel,
       int olabel,
       float weight = 0.0);
@@ -330,43 +330,43 @@ class Graph {
     return sharedGraph_->accept;
   };
   /** Check if the `i`-th node is a start node. */
-  bool isStart(int i) const {
+  bool isStart(size_t i) const {
     return node(i).start;
   };
   /** Check if the `i`-th node is an accepting node. */
-  bool isAccept(int i) const {
+  bool isAccept(size_t i) const {
     return node(i).accept;
   };
   /** Make the the `i`-th node an accepting node. */
-  void makeAccept(int i) {
+  void makeAccept(size_t i) {
     auto& n = node(i);
     if (!n.accept) {
-      sharedGraph_->accept.push_back(i);
+      sharedGraph_->accept.push_back(static_cast<int>(i));
       n.accept = true;
     }
   };
   /** The number of outgoing arcs from the `i`-th node. */
-  size_t numOut(int i) const {
+  size_t numOut(size_t i) const {
     return node(i).out.size();
   }
   /** Get the indices of outgoing arcs from the `i`-th node. */
-  const std::vector<int>& out(int i) const {
+  const std::vector<int>& out(size_t i) const {
     return node(i).out;
   }
   /** Get the index of the `j`-th outgoing arc from the `i`-th node. */
-  int out(int i, int j) const {
+  int out(size_t i, size_t j) const {
     return node(i).out[j];
   }
   /** The number of incoming arcs to the `i`-th node. */
-  size_t numIn(int i) const {
+  size_t numIn(size_t i) const {
     return node(i).in.size();
   }
   /** Get the indices of incoming arcs to the `i`-th node. */
-  const std::vector<int>& in(int i) const {
+  const std::vector<int>& in(size_t i) const {
     return node(i).in;
   }
   /** Get the index of the `j`-th incoming arc to the `i`-th node. */
-  int in(int i, int j) const {
+  size_t in(size_t i, size_t j) const {
     return node(i).in[j];
   }
 
@@ -377,33 +377,33 @@ class Graph {
     */
 
   /** The destination node of the `i`-th arc. */
-  int srcNode(int i) const {
+  int srcNode(size_t i) const {
     return arc(i).srcNode;
   }
   /** The source node of the `i`-th arc. */
-  int dstNode(int i) const {
+  int dstNode(size_t i) const {
     return arc(i).dstNode;
   }
   /** The label of the `i`-th arc (use this for acceptors). */
-  int label(int i) const {
+  int label(size_t i) const {
     return arc(i).ilabel;
   }
   /** The input label of the `i`-th arc. */
-  int ilabel(int i) const {
+  int ilabel(size_t i) const {
     return arc(i).ilabel;
   }
   /** The output label of the `i`-th arc. */
-  int olabel(int i) const {
+  int olabel(size_t i) const {
     return arc(i).olabel;
   }
 
   /** The weight of the `i`-th arc. */
-  float weight(int i) const {
+  float weight(size_t i) const {
     assert(sharedWeights_ != nullptr);
     return (*sharedWeights_)[i];
   }
   /** Set the weight of the `i`-th arc. */
-  void setWeight(int i, float weight) {
+  void setWeight(size_t i, float weight) {
     assert(sharedWeights_ != nullptr);
     (*sharedWeights_)[i] = weight;
   }
@@ -411,8 +411,8 @@ class Graph {
 
  private:
   // Attempt to keep code like `g.addArc(n1, n2, 0, 2.0)` from compiling
-  int addArc(int srcNode, int dstNode, int label, float) = delete;
-  int addArc(int srcNode, int dstNode, int label, double) = delete;
+  size_t addArc(size_t srcNode, size_t dstNode, int label, float) = delete;
+  size_t addArc(size_t srcNode, size_t dstNode, int label, double) = delete;
 
   const Node& node(size_t i) const {
     // NB: assert gets stripped at in release mode
